move controllers into pkg/
